Create README.md
The Hospital Management System (HMS) is designed for any Hospital to replace their existing manual, paper - based system. This system targets to provide complete solution for Hospital and Health care services. This system can be used in any Hospital, Clinic, Diagnostics or Pathology labs for maintaining patient details and their test results. It integrates the entire resources of a hospital into One Integrated Software Application. It allows user to easily understand and edit the program. This system is focusing on the maintaining records of patients and their details in a computerized system rather than time taking and cumbersome manual system.
This Project is a fine thought to make the complex

procedure of the Hospital management system to an easy manner which is systematic, modular designed,


selective menu based user display. The HMS (Hospital Management System) project is developed with the hope of us to bring the friendliest and most efficient way of management in order to alleviate the workload of hospital staff.